Extracted many extension classes from FakeExtensions. For #206.
