+++ Source/FakeItEasy-SL/FakeItEasy-SL.csproj
@@ -98,11 +98,23 @@
     <Compile Include="..\FakeItEasy\ArgumentConstraintExtensions.cs">
       <Link>ArgumentConstraintExtensions.cs</Link>
     </Compile>
+    <Compile Include="..\FakeItEasy\WhereExtensions.cs">
+      <Link>WhereExtensions.cs</Link>
+    </Compile>
+    <Compile Include="..\FakeItEasy\WithAnyArgumentsExtensions.cs">
+      <Link>WithAnyArgumentsExtensions.cs</Link>
+    </Compile>
+    <Compile Include="..\FakeItEasy\WriteExtensions.cs">
+      <Link>WriteExtensions.cs</Link>
+    </Compile>
     <Compile Include="..\FakeItEasy\ArgumentValueFormatter.cs">
       <Link>ArgumentValueFormatter.cs</Link>
     </Compile>
     <Compile Include="..\FakeItEasy\CommonExtensions.cs">
       <Link>CommonExtensions.cs</Link>
+    </Compile>
+    <Compile Include="..\FakeItEasy\MatchingExtensions.cs">
+      <Link>MatchingExtensions.cs</Link>
     </Compile>
     <Compile Include="..\FakeItEasy\Configuration\AnyCallCallRule.cs">
       <Link>Configuration\AnyCallCallRule.cs</Link>
@@ -527,11 +539,11 @@
     <Compile Include="..\FakeItEasy\FakeConfigurator.cs">
       <Link>FakeConfigurator.cs</Link>
     </Compile>
-    <Compile Include="..\FakeItEasy\FakeExtensions.cs">
-      <Link>FakeExtensions.cs</Link>
-    </Compile>
     <Compile Include="..\FakeItEasy\FakeFacade.cs">
       <Link>FakeFacade.cs</Link>
+    </Compile>
+    <Compile Include="..\FakeItEasy\GetArgumentExtensions.cs">
+      <Link>GetArgumentExtensions.cs</Link>
     </Compile>
     <Compile Include="..\FakeItEasy\Guard.cs">
       <Link>Guard.cs</Link>
@@ -557,6 +569,9 @@
     <Compile Include="..\FakeItEasy\IFileSystem.cs">
       <Link>IFileSystem.cs</Link>
     </Compile>
+    <Compile Include="..\FakeItEasy\InvokesExtensions.cs">
+      <Link>InvokesExtensions.cs</Link>
+    </Compile>
     <Compile Include="..\FakeItEasy\IoC\DictionaryContainer.cs">
       <Link>IoC\DictionaryContainer.cs</Link>
     </Compile>
@@ -572,8 +587,17 @@
     <Compile Include="..\FakeItEasy\MustBeImplementedException.cs">
       <Link>MustBeImplementedException.cs</Link>
     </Compile>
+    <Compile Include="..\FakeItEasy\MustHaveHappenedExtensions.cs">
+      <Link>MustHaveHappenedExtensions.cs</Link>
+    </Compile>
+    <Compile Include="..\FakeItEasy\MustNotHaveHappenedExtensions.cs">
+      <Link>MustNotHaveHappenedExtensions.cs</Link>
+    </Compile>
     <Compile Include="..\FakeItEasy\NextCall.cs">
       <Link>NextCall.cs</Link>
+    </Compile>
+    <Compile Include="..\FakeItEasy\OnceExtensions.cs">
+      <Link>OnceExtensions.cs</Link>
     </Compile>
     <Compile Include="..\FakeItEasy\OrderedAssertion.cs">
       <Link>OrderedAssertion.cs</Link>
@@ -595,6 +619,15 @@
     </Compile>
     <Compile Include="..\FakeItEasy\Repeated.cs">
       <Link>Repeated.cs</Link>
+    </Compile>
+    <Compile Include="..\FakeItEasy\ReturnsExtensions.cs">
+      <Link>ReturnsExtensions.cs</Link>
+    </Compile>
+    <Compile Include="..\FakeItEasy\ReturnsLazilyExtensions.cs">
+      <Link>ReturnsLazilyExtensions.cs</Link>
+    </Compile>
+    <Compile Include="..\FakeItEasy\ReturnsNextFromSequenceExtensions.cs">
+      <Link>ReturnsNextFromSequenceExtensions.cs</Link>
     </Compile>
     <Compile Include="..\FakeItEasy\RootModule.cs">
       <Link>RootModule.cs</Link>
@@ -626,11 +659,23 @@
     <Compile Include="..\FakeItEasy\SmellyAttribute.cs">
       <Link>SmellyAttribute.cs</Link>
     </Compile>
+    <Compile Include="..\FakeItEasy\StrictExtensions.cs">
+      <Link>StrictExtensions.cs</Link>
+    </Compile>
     <Compile Include="..\FakeItEasy\StringBuilderOutputWriter.cs">
       <Link>StringBuilderOutputWriter.cs</Link>
     </Compile>
+    <Compile Include="..\FakeItEasy\ThrowsExtensions.cs">
+      <Link>ThrowsExtensions.cs</Link>
+    </Compile>
+    <Compile Include="..\FakeItEasy\TwiceExtensions.cs">
+      <Link>TwiceExtensions.cs</Link>
+    </Compile>
     <Compile Include="..\FakeItEasy\UnderTestAttribute.cs">
       <Link>UnderTestAttribute.cs</Link>
+    </Compile>
+    <Compile Include="..\FakeItEasy\ValueProducerSignatureHelper.cs">
+      <Link>ValueProducerSignatureHelper.cs</Link>
     </Compile>
     <Compile Include="..\SharedAssemblyInfo.cs">
       <Link>SharedAssemblyInfo.cs</Link>
