+++ Source/FakeItEasy.Net35/FakeItEasy.Net35.csproj
@@ -508,11 +508,11 @@
     <Compile Include="..\FakeItEasy\FakeConfigurator.cs">
       <Link>FakeConfigurator.cs</Link>
     </Compile>
-    <Compile Include="..\FakeItEasy\FakeExtensions.cs">
-      <Link>FakeExtensions.cs</Link>
-    </Compile>
     <Compile Include="..\FakeItEasy\FakeFacade.cs">
       <Link>FakeFacade.cs</Link>
+    </Compile>
+    <Compile Include="..\FakeItEasy\GetArgumentExtensions.cs">
+      <Link>GetArgumentExtensions.cs</Link>
     </Compile>
     <Compile Include="..\FakeItEasy\Guard.cs">
       <Link>Guard.cs</Link>
@@ -541,6 +541,9 @@
     <Compile Include="..\FakeItEasy\ImportsModule.cs">
       <Link>ImportsModule.cs</Link>
     </Compile>
+    <Compile Include="..\FakeItEasy\InvokesExtensions.cs">
+      <Link>InvokesExtensions.cs</Link>
+    </Compile>
     <Compile Include="..\FakeItEasy\IoC\DictionaryContainer.cs">
       <Link>IoC\DictionaryContainer.cs</Link>
     </Compile>
@@ -553,11 +556,23 @@
     <Compile Include="..\FakeItEasy\IRepeatSpecification.cs">
       <Link>IRepeatSpecification.cs</Link>
     </Compile>
+    <Compile Include="..\FakeItEasy\MatchingExtensions.cs">
+      <Link>MatchingExtensions.cs</Link>
+    </Compile>
     <Compile Include="..\FakeItEasy\MustBeImplementedException.cs">
       <Link>MustBeImplementedException.cs</Link>
     </Compile>
+    <Compile Include="..\FakeItEasy\MustHaveHappenedExtensions.cs">
+      <Link>MustHaveHappenedExtensions.cs</Link>
+    </Compile>
+    <Compile Include="..\FakeItEasy\MustNotHaveHappenedExtensions.cs">
+      <Link>MustNotHaveHappenedExtensions.cs</Link>
+    </Compile>
     <Compile Include="..\FakeItEasy\NextCall.cs">
       <Link>NextCall.cs</Link>
+    </Compile>
+    <Compile Include="..\FakeItEasy\OnceExtensions.cs">
+      <Link>OnceExtensions.cs</Link>
     </Compile>
     <Compile Include="..\FakeItEasy\OrderedAssertion.cs">
       <Link>OrderedAssertion.cs</Link>
@@ -579,6 +594,15 @@
     </Compile>
     <Compile Include="..\FakeItEasy\Repeated.cs">
       <Link>Repeated.cs</Link>
+    </Compile>
+    <Compile Include="..\FakeItEasy\ReturnsExtensions.cs">
+      <Link>ReturnsExtensions.cs</Link>
+    </Compile>
+    <Compile Include="..\FakeItEasy\ReturnsLazilyExtensions.cs">
+      <Link>ReturnsLazilyExtensions.cs</Link>
+    </Compile>
+    <Compile Include="..\FakeItEasy\ReturnsNextFromSequenceExtensions.cs">
+      <Link>ReturnsNextFromSequenceExtensions.cs</Link>
     </Compile>
     <Compile Include="..\FakeItEasy\RootModule.cs">
       <Link>RootModule.cs</Link>
@@ -610,11 +634,32 @@
     <Compile Include="..\FakeItEasy\SmellyAttribute.cs">
       <Link>SmellyAttribute.cs</Link>
     </Compile>
+    <Compile Include="..\FakeItEasy\StrictExtensions.cs">
+      <Link>StrictExtensions.cs</Link>
+    </Compile>
     <Compile Include="..\FakeItEasy\StringBuilderOutputWriter.cs">
       <Link>StringBuilderOutputWriter.cs</Link>
     </Compile>
+    <Compile Include="..\FakeItEasy\ThrowsExtensions.cs">
+      <Link>ThrowsExtensions.cs</Link>
+    </Compile>
+    <Compile Include="..\FakeItEasy\TwiceExtensions.cs">
+      <Link>TwiceExtensions.cs</Link>
+    </Compile>
     <Compile Include="..\FakeItEasy\UnderTestAttribute.cs">
       <Link>UnderTestAttribute.cs</Link>
+    </Compile>
+    <Compile Include="..\FakeItEasy\ValueProducerSignatureHelper.cs">
+      <Link>ValueProducerSignatureHelper.cs</Link>
+    </Compile>
+    <Compile Include="..\FakeItEasy\WhereExtensions.cs">
+      <Link>WhereExtensions.cs</Link>
+    </Compile>
+    <Compile Include="..\FakeItEasy\WithAnyArgumentsExtensions.cs">
+      <Link>WithAnyArgumentsExtensions.cs</Link>
+    </Compile>
+    <Compile Include="..\FakeItEasy\WriteExtensions.cs">
+      <Link>WriteExtensions.cs</Link>
     </Compile>
     <Compile Include="..\InternalsVisibleTo-Signed.cs">
       <Link>InternalsVisibleTo-Signed.cs</Link>
