+++ Source/FakeItEasy/ReturnsExtensions.cs
@@ -1,0 +1,24 @@
+namespace FakeItEasy
+{
+    using FakeItEasy.Configuration;
+
+    /// <summary>
+    /// Provides the Returns extension method for specifying return values of fake object calls.
+    /// </summary>
+    public static class ReturnsExtensions
+    {
+        /// <summary>
+        /// Specifies the value to return when the configured call is made.
+        /// </summary>
+        /// <typeparam name="T">The type of the return value.</typeparam>
+        /// <param name="configuration">The configuration to extend.</param>
+        /// <param name="value">The value to return.</param>
+        /// <returns>A configuration object.</returns>
+        public static IAfterCallSpecifiedWithOutAndRefParametersConfiguration Returns<T>(this IReturnValueConfiguration<T> configuration, T value)
+        {
+            Guard.AgainstNull(configuration, "configuration");
+
+            return configuration.ReturnsLazily(x => value);
+        }
+    }
+}
