+++ Source/FakeItEasy/ReturnsLazilyExtensions.cs
@@ -1,0 +1,133 @@
+namespace FakeItEasy
+{
+    using System;
+
+    using FakeItEasy.Configuration;
+
+    /// <summary>
+    /// Provides the ReturnsLazily extension methods for specifying return values of fake object calls.
+    /// </summary>
+    public static class ReturnsLazilyExtensions
+    {
+        private const string NameOfReturnsLazilyFeature = "returns lazily";
+
+        /// <summary>
+        /// Specifies a function used to produce a return value when the configured call is made.
+        /// The function will be called each time this call is made and can return different values
+        /// each time.
+        /// </summary>
+        /// <typeparam name="T">The type of the return value.</typeparam>
+        /// <param name="configuration">The configuration to extend.</param>
+        /// <param name="valueProducer">A function that produces the return value.</param>
+        /// <returns>A configuration object.</returns>
+        public static IAfterCallSpecifiedWithOutAndRefParametersConfiguration ReturnsLazily<T>(this IReturnValueConfiguration<T> configuration, Func<T> valueProducer)
+        {
+            Guard.AgainstNull(configuration, "configuration");
+            Guard.AgainstNull(valueProducer, "valueProducer");
+
+            return configuration.ReturnsLazily(x => valueProducer());
+        }
+
+        /// <summary>
+        /// Specifies a function used to produce a return value when the configured call is made.
+        /// The function will be called each time this call is made and can return different values
+        /// each time.
+        /// </summary>
+        /// <typeparam name="TReturnType">The type of the return value.</typeparam>
+        /// <typeparam name="T1">Type of the first argument of the faked method call.</typeparam>
+        /// <param name="configuration">The configuration to extend.</param>
+        /// <param name="valueProducer">A function that produces the return value.</param>
+        /// <returns>A configuration object.</returns>
+        /// <exception cref="FakeConfigurationException">The signatures of the faked method and the <paramref name="valueProducer" /> do not match.</exception>
+        public static IAfterCallSpecifiedWithOutAndRefParametersConfiguration
+            ReturnsLazily<TReturnType, T1>(this IReturnValueConfiguration<TReturnType> configuration, Func<T1, TReturnType> valueProducer)
+        {
+            Guard.AgainstNull(configuration, "configuration");
+
+            return configuration.ReturnsLazily(call =>
+                {
+                    ValueProducerSignatureHelper.AssertThatValueProducerSignatureSatisfiesCallSignature(call.Method, valueProducer.Method, NameOfReturnsLazilyFeature);
+
+                    return valueProducer(call.GetArgument<T1>(0));
+                });
+        }
+
+        /// <summary>
+        /// Specifies a function used to produce a return value when the configured call is made.
+        /// The function will be called each time this call is made and can return different values
+        /// each time.
+        /// </summary>
+        /// <param name="configuration">The configuration to extend.</param>
+        /// <param name="valueProducer">A function that produces the return value.</param>
+        /// <typeparam name="TReturnType">The type of the return value.</typeparam>
+        /// <typeparam name="T1">Type of the first argument of the faked method call.</typeparam>
+        /// <typeparam name="T2">Type of the second argument of the faked method call.</typeparam>
+        /// <returns>A configuration object.</returns>
+        /// <exception cref="FakeConfigurationException">The signatures of the faked method and the <paramref name="valueProducer"/> do not match.</exception>
+        public static IAfterCallSpecifiedWithOutAndRefParametersConfiguration
+            ReturnsLazily<TReturnType, T1, T2>(this IReturnValueConfiguration<TReturnType> configuration, Func<T1, T2, TReturnType> valueProducer)
+        {
+            Guard.AgainstNull(configuration, "configuration");
+
+            return configuration.ReturnsLazily(call =>
+                {
+                    ValueProducerSignatureHelper.AssertThatValueProducerSignatureSatisfiesCallSignature(call.Method, valueProducer.Method, NameOfReturnsLazilyFeature);
+
+                    return valueProducer(call.GetArgument<T1>(0), call.GetArgument<T2>(1));
+                });
+        }
+
+        /// <summary>
+        /// Specifies a function used to produce a return value when the configured call is made.
+        /// The function will be called each time this call is made and can return different values
+        /// each time.
+        /// </summary>
+        /// <param name="configuration">The configuration to extend.</param>
+        /// <param name="valueProducer">A function that produces the return value.</param>
+        /// <typeparam name="TReturnType">The type of the return value.</typeparam>
+        /// <typeparam name="T1">Type of the first argument of the faked method call.</typeparam>
+        /// <typeparam name="T2">Type of the second argument of the faked method call.</typeparam>
+        /// <typeparam name="T3">Type of the third argument of the faked method call.</typeparam>
+        /// <returns>A configuration object.</returns>
+        /// <exception cref="FakeConfigurationException">The signatures of the faked method and the <paramref name="valueProducer"/> do not match.</exception>
+        public static IAfterCallSpecifiedWithOutAndRefParametersConfiguration
+            ReturnsLazily<TReturnType, T1, T2, T3>(this IReturnValueConfiguration<TReturnType> configuration, Func<T1, T2, T3, TReturnType> valueProducer)
+        {
+            Guard.AgainstNull(configuration, "configuration");
+
+            return configuration.ReturnsLazily(call =>
+                {
+                    ValueProducerSignatureHelper.AssertThatValueProducerSignatureSatisfiesCallSignature(call.Method, valueProducer.Method, NameOfReturnsLazilyFeature);
+
+                    return valueProducer(call.GetArgument<T1>(0), call.GetArgument<T2>(1), call.GetArgument<T3>(2));
+                });
+        }
+
+        /// <summary>
+        /// Specifies a function used to produce a return value when the configured call is made.
+        /// The function will be called each time this call is made and can return different values
+        /// each time.
+        /// </summary>
+        /// <param name="configuration">The configuration to extend.</param>
+        /// <param name="valueProducer">A function that produces the return value.</param>
+        /// <typeparam name="TReturnType">The type of the return value.</typeparam>
+        /// <typeparam name="T1">Type of the first argument of the faked method call.</typeparam>
+        /// <typeparam name="T2">Type of the second argument of the faked method call.</typeparam>
+        /// <typeparam name="T3">Type of the third argument of the faked method call.</typeparam>
+        /// <typeparam name="T4">Type of the fourth argument of the faked method call.</typeparam>
+        /// <returns>A configuration object.</returns>
+        /// <exception cref="FakeConfigurationException">The signatures of the faked method and the <paramref name="valueProducer"/> do not match.</exception>
+        public static IAfterCallSpecifiedWithOutAndRefParametersConfiguration
+            ReturnsLazily<TReturnType, T1, T2, T3, T4>(this IReturnValueConfiguration<TReturnType> configuration, Func<T1, T2, T3, T4, TReturnType> valueProducer)
+        {
+            Guard.AgainstNull(configuration, "configuration");
+
+            return configuration.ReturnsLazily(call =>
+                {
+                    ValueProducerSignatureHelper.AssertThatValueProducerSignatureSatisfiesCallSignature(call.Method, valueProducer.Method, NameOfReturnsLazilyFeature);
+
+                    return valueProducer(call.GetArgument<T1>(0), call.GetArgument<T2>(1), call.GetArgument<T3>(2), call.GetArgument<T4>(3));
+                });
+        }
+    }
+}
